Form dashboard shiam

+++ Backend/controllers/files.controller.js
@@ -77,6 +77,9 @@ export const downloadFile = async (req, res) => {
 export const deleteFile = async (req, res) => {
   try {
     const fileId = req.params.id;
+    if (!fileId) {
+      return res.status(400).json({ message: "File ID not found" });
+    }
 
     // Find the file by ID and delete it
     const deletedFile = await File.findByIdAndDelete(fileId);
@@ -84,7 +87,7 @@ export const deleteFile = async (req, res) => {
       return res.status(404).json({ message: 'File not found' });
     }
 
-    res.json({ message: 'File deleted successfully' });
+    res.status(200).json({ message: 'File deleted successfully' });
   } catch (error) {
     res.status(500).json({ message: 'Error deleting file', error: error.message });
   }
@@ -116,13 +119,16 @@ export const listFilesByOrderID = async (req, res) => {
 // Controller to generate a shareable link
 export const generateShareableLink = (req, res) => {
   try {
-    const { orderID } = req.body;
+    const { orderID } = req.body;    
 
     if (!orderID) {
       return res.status(400).json({ message: 'Order ID is required' });
     }
 
-    const shareableLink = `${req.protocol}://${req.get('host')}/api/files/access/${orderID}`;
+    // const shareableLink = `${req.protocol}://${req.get('host')}/api/files/access/${orderID}`;
+    const shareableLink = `${orderID}`;
+
+    
     res.status(200).json({ message: 'Shareable link generated successfully', shareableLink });
   } catch (error) {
     res.status(500).json({ message: 'Error generating shareable link', error: error.message });
